Update all links to work with nested routing
Not all current links were functional or pointing to the expected
location. Updated paths and controllers to behave as expected. Getting
delete links to work required adding rails-ujs for JavaScript via the
asset-pipeline and including it in the app layout file. Feature tests
were added for all happy-path actions to build a regression suite.